# Add optimized version of mpo construction for approximate tensor